refactor: remove setTimeout in order to have nicer traceback
